+++ packages/xchain-utxo-providers/src/providers/index.ts
@@ -32,3 +32,5 @@ export * from './blockcypher/blockcypher-data-provider'
 export { BlockcypherNetwork } from './blockcypher/blockcypher-api-types'
 
 export { BitgoProvider } from './bitgo/bitgo-data-provider'
+
+export { NownodesProvider } from './nownodes/nownodes-data-provider'