+++ packages/xchain-zcash/src/index.ts
@@ -1,0 +1,14 @@
+// Export the 'ClientKeystore' class as 'Client'
+export { ClientKeystore as Client } from './clientKeystore'
+
+// Export the 'ClientLedger' class
+export { ClientLedger } from './clientLedger'
+
+// Export the 'defaultBTCParams' constant from the 'client' file
+export { defaultZECParams } from './client'
+
+// Export all constants from the 'const' file
+export * from './const'
+
+// Export the 'getPrefix' and 'validateAddress' functions from the 'utils' file
+export { getPrefix, validateAddress } from './utils'